Updated manual token refresh workflow

+++ manual_token_refresh.py
@@ -1,7 +1,10 @@
 import os
 from pydrive2.auth import GoogleAuth
 
-os.remove("credentials.json")
+# Remove existing credentials if it exists
+if os.path.exists("credentials.json"):
+    os.remove("credentials.json")
 
+# Reauth with Google
 gauth = GoogleAuth()
-gauth.CommandLineAuth()
+gauth.LocalWebserverAuth()
